pyproject.toml
[tool.poetry]
name = "Z3R0"
version = "3.4.6"
description = "A multi-purpose open source discord bot"
authors = ["null2264"]
license = "MPL-2.0"
packages = [
  {include = "main", from = "src"}
]

[tool.poetry.scripts]
bot = "main.__main__:main"

[tool.poetry.dependencies]
aiohttp = "3.7.3"
python = "^3.9"
beautifulsoup4 = "^4.9.3"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", rev = "b20ab5ae5d56a87539c8399acd53f96c927331e9"}
discord-ext-menus = "^1.1"
fuzzywuzzy = "0.18.0"
humanize = "3.7.0"
jishaku = "^2.2.0"
Pillow = "^8.3.1"
python-dateutil = "^2.8.2"
python-Levenshtein = "^0.12.2"
pytz = "^2021.1"
pyparsing = "^2.4.7"
requests = "^2.26.0"
tagscriptengine = {git = "https://github.com/ZiRO-Bot/TagScript", rev = "99a7999d61d891b0ed75ad2f2767eb49a7a9abc6"}
uvloop = {version = "^0.15.3", optional = true}
typing-extensions = "^3.10.0"
asyncpg = {version = "^0.23.0", optional = true}
aiomysql = {version = "^0.0.21", optional = true}
aiosqlite = ">=0.16.0, <0.18.0"
asyncmy = { version = "*", optional = true }
PyNaCl = {version = ">=1.3.0,<1.5", optional = true}
tortoise-orm = "^0.17.7"
orjson = {version = "^3.6.3", optional = true}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
pre-commit = "^2.13.0"
pyproject-flake8 = "^0.0.1-alpha.2"
aerich = "^0.5.8"

[tool.poetry.extras]
postgresql = ["asyncpg"]
mysql = ["aiomysql"]
"mysql+asyncmy" = ["asyncmy"]
voice = ["PyNaCl"]
speedup = ["uvloop", "orjson"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
skip-string-normalization = true

[tool.flake8]
max-line-length = 125
select = "C,E,F,W,B,B950"
extend-ignore = "E203,E501"

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
